#12 - list group members and expenses

+++ backend/app/routers/group_router.py
@@ -6,7 +6,7 @@ from sqlmodel import Session
 
 from app.core.db import get_db
 from app.models import Group, User
-from app.schemas.group_schemas import GroupCreate
+from app.schemas.group_schemas import GroupCreate, GroupView
 from app.services import group_service
 from app.services.auth_service import get_current_user
 
@@ -18,7 +18,7 @@ def create_group(group: GroupCreate, user: User = Depends(get_current_user), db:
     return group_service.create_group(db, group, user.username)
 
 
-@group_router.get("/groups/{group_id}", response_model=Group)
+@group_router.get("/groups/{group_id}", response_model=GroupView)
 def get_group(group_id: UUID, db: Session = Depends(get_db)):
     return group_service.get_group(db, group_id)
 
@@ -26,7 +26,9 @@ def get_group(group_id: UUID, db: Session = Depends(get_db)):
 @group_router.delete("/groups/{group_id}")
 def delete_group(group_id: UUID, user: User = Depends(get_current_user), db: Session = Depends(get_db)):
     # if user does not belong to the group, raise 403
-    group = group_service.get_group(db, group_id)
+    group = db.get(Group, group_id)
+    if group is None:
+        raise HTTPException(status_code=404, detail="Group not found")
     if group not in user.groups:
         raise HTTPException(status_code=403, detail="User does not belong to the group")
     group_service.delete_group(db, group_id)
@@ -38,7 +40,9 @@ def add_user_to_group(
     group_id: str, username: str, user: User = Depends(get_current_user), db: Session = Depends(get_db)
 ):
     # if user does not belong to the group, raise 403
-    group = group_service.get_group(db, group_id)
+    group = db.get(Group, group_id)
+    if group is None:
+        raise HTTPException(status_code=404, detail="Group not found")
     if group not in user.groups:
         raise HTTPException(status_code=403, detail="User does not belong to the group")
     group_service.add_user_to_group(db, username, group_id)
@@ -49,7 +53,9 @@ def add_user_to_group(
 def remove_user_from_group(
     group_id: str, username: str, user: User = Depends(get_current_user), db: Session = Depends(get_db)
 ):
-    group = group_service.get_group(db, group_id)
+    group = db.get(Group, group_id)
+    if group is None:
+        raise HTTPException(status_code=404, detail="Group not found")
     if group not in user.groups:
         raise HTTPException(status_code=403, detail="User does not belong to the group")
     group_service.remove_user_from_group(db, username, group_id)
